Escape control characters; fix CRM issues
Since ESC is now preserved for unrecognised sequences, I feel able to
use it to escape control characters in order to display them (e.g.
"\e\e" will display a single escape).

Control mode would not display the recognised controls (i.e. BEL, SO and
SI) and would not work if it was turned off in the same string.