Add DECOM, DECSTBM, SD and SU
Margins always use the window, never the buffer; the exception is
`\e[+r` which will remove the margins ('\e[r' will set the margins to
the window).  With the margins set, the window will not scroll when the
cursor is at the bottom, outside of the margins.  In this case, the
display is not quite right, since I copy the line from another buffer,
rather than actually overwrite the existing line.