Set margins per console; fix windowless processes
I had intended for margins to be set per process, but during testing I
was wondering why they weren't working, so per console it is.

In changing the above, I realised that `pState` may stay `NULL` when a
process is created without a window.  Initialise it to a default state,
updating it when the window becomes available.