Fix CR and CRM issues
CR and BS would still be processed during CRM.	Fix all the partial CRM
sequences followed by a complete CRM sequence.

It's possible for some text files to have `\r\r\n` endings (Microsoft
converts LF to CRLF even when the CR is already there).  Collapse
multiple CRs to a single CR to catch this.  It is also necessary to
always delay flushing if the last character is CR, to see if the next
character is CR or LF.